Refactor constants.
