+++ src/main/java/constants/TravisConstants.java
@@ -1,0 +1,12 @@
+package constants;
+
+public final class TravisConstants {
+    public static final String GREETING = "Hello! I'm Travis!\nWhat can I do for you?";
+    public static final String FAREWELL = "Bye. Hope to see you again soon!";
+    public static final String HORIZONTAL_LINE = "____________________________________________________________";
+    public static final String TOTAL_TASKS = "Now you have %d task(s) in your list.";
+    public static final String NEW_TASK = "A new task has been added: %s\n";
+    public static final String DELETED_TASK = "Noted. I have deleted the following task: %s\n";
+    public static final String MARKED_AS_DONE = "Nice! I've marked this task as done: %s";
+    public static final String MARKED_AS_NOT_DONE = "Ok, I've marked this task as not done yet: %s";
+}
